refactor: move upstream to it's own file
