+++ src/blueprint/from_config/mod.rs
@@ -2,8 +2,10 @@ mod from_config;
 mod server;
 mod graphql;
 mod types;
+mod upstream;
 
 pub use server::*;
 pub use from_config::*;
 pub use graphql::*;
 pub use types::*;
+pub use upstream::*;
